Function adapton::collections::list_of_vec

pub fn list_of_vec<X: 'static + Clone, L: ListIntro<X>>(v: &Vec<NameElse<X>>)
                                                        -> L

Constructs a linked list that consists of elements and names, as given by the input vector (in that order). Not incremental; used only for setting up inputs for tests.