Function adapton::collections::raz::zip_insert

pub fn zip_insert<X: Eq + Clone + Hash + Debug>(z: Zip<X>,
                                                d: Dir,
                                                x: X,
                                                p: Option<Punc>)
                                                -> Zip<X>

Transform the zipper, inserting element x in the given direction d. The optional punctuation p follows the inserted element, in the given direction.