Module adapton::engine::reflect

Structs

CompNode
Reflected version of engine::CompNode

DCG
Reflected version of engine::DCG

Frame
Reflected version of engine::Frame

Loc
Reflected version of engine::Loc

Pred
Reflected version of engine::Pred

RefNode
Reflected version of engine::MutNode

Succ
Reflected version of engine::Succ

Enums

ArtContent
The content of an articulation: Either a cell holding a value, or a thunk that has optionally produced a value.

Effect
Reflected version of engine::Effect

GraphNode
Reflected version of engine::GraphNode

Val
Reflected value; Gives a syntax for inductive data type constructors (Constr), named articulations (Art) and primitive data (`Data).

Traits

Data
Reflected data